### `requirements.txt`.txt
### `requirements.txt`

```plaintext
numpy
pandas
umap-learn
matplotlib